Add checks to prevent orphaned records

+++ app/controllers/admin/LocationsController.php
@@ -175,11 +175,21 @@ class LocationsController extends AdminController {
 			return Redirect::to('admin/settings/locations')->with('error', Lang::get('admin/locations/message.not_found'));
 		}
 
-		// Delete the location
-		$location->delete();
 
-		// Redirect to the locations management page
-		return Redirect::to('admin/settings/locations')->with('success', Lang::get('admin/locations/message.delete.success'));
+		if ($location->has_users() > 0) {
+
+			// Redirect to the asset management page
+			return Redirect::to('admin/settings/locations')->with('error', Lang::get('admin/locations/message.assoc_users'));
+		} else {
+
+			$location->delete();
+
+			// Redirect to the locations management page
+			return Redirect::to('admin/settings/locations')->with('success', Lang::get('admin/locations/message.delete.success'));
+		}
+
+
+
 	}
